+++ main.tf
@@ -6,7 +6,8 @@ resource "aws_api_gateway_rest_api" "this" {
   tags              = var.tags
 
   endpoint_configuration {
-    types = [var.endpoint_type]
+    types                        = [var.endpoint_type]
+    disable_execute_api_endpoint = var.disable_execute_api_endpoint
   }
 }